requirements.txt
# PyTorch y aceleradores
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0

# Transformers y NLP
transformers==4.44.0
tokenizers==0.19.1
sentencepiece==0.2.0
sacremoses==0.1.1

# Métricas
sacrebleu==2.4.2
nltk==3.8.1
datasets==2.20.0

# Visualización y análisis
matplotlib==3.9.0
seaborn==0.13.2
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4

# Testing
pytest==8.2.2
pytest-cov==5.0.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.1

# Notebooks
jupyter==1.0.0
ipykernel==6.29.4
ipywidgets==8.1.3

# Profiling
py-cpuinfo==9.0.0
psutil==6.0.0
